+++ src/App.css
@@ -7,3 +7,24 @@ body {
   padding: 2rem;
   text-align: center;
 }
+.formContainer{
+margin: 30px 0;
+display: flex;
+align-items: center;
+}
+.formContainer :is(label, input, button) {
+  margin: 10px 20px 0px -0px;
+  padding: 7px;
+  width: 100%;
+}
+.formContainer input {
+  width: 200px;
+}
+ul {
+  padding: 0;
+  margin: 0;
+  list-style-position:inside;
+}
+li {
+  padding: 2px;
+}
